Attempt to make bin removal and info lookup more robust.
Sometimes deleting large query bins can fail or timeout, removing
the tweet tables but not the entry from the bins table. This
results in the app crashing when accessing the capture page.

This change attempts to prevent this by a) putting deletion
operations in a single transaction and b) catching errors when
retrieving bin info. The latter change is required because MySQL
does not do schema changes such as table deletion in transactions
so the problem with incomplete bin removal can still occur.